requirements.txt
alembic==1.13.1
Babel==2.14.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
